perf: Shrink memory allocation for images
Images takes a lot of memory because they are converted to Bitmap. Solution can be to use DecodePixelWidth/Height in Image/Image.Source/BitmapImage in XAML. Unfortunately it does not support binding UriSource to ViewModel. In order to cope with that Uri-to-Bitmap converter is implemented.
Link to documentation: https://docs.microsoft.com/en-us/dotnet/desktop/wpf/graphics-multimedia/imaging-overview